Change the types to int.

+++ Cryptography/ShiftCipher.java
@@ -1,34 +1,34 @@
 public class ShiftCipher {
-    private final short NUM_CHARS = 256;
-    private short key;
+    private final int NUM_CHARS = 256;
+    private int key;
 
-    public ShiftCipher(short key) {
+    public ShiftCipher(int key) {
         if (key < 0 || key >= NUM_CHARS) {
             throw new RuntimeException("Choose a key in the interval [0 ; 256)");
         }
         this.setKey(key);
     }
 
-    public void setKey(short key) {
+    public void setKey(int key) {
         this.key = key;
     }
 
-    public short getKey() {
+    public int getKey() {
         return key;
     }
 
-    public short[] encrypt(String data) {
-        short[] dataEncrypted = new short[data.length()];
+    public int[] encrypt(String data) {
+        int[] dataEncrypted = new int[data.length()];
         for (int i = 0; i < data.length(); i++) {
-            dataEncrypted[i] = (short) ((data.charAt(i) + key) % NUM_CHARS);
+            dataEncrypted[i] = (int) ((data.charAt(i) + key) % NUM_CHARS);
         }
         return dataEncrypted;
     }
 
-    public String decrypt(short[] data) {
+    public String decrypt(int[] data) {
         String dataDecrypted = "";
         for (int i = 0; i < data.length; i++) {
-            dataDecrypted += (short) ((data[i] - key) % NUM_CHARS);
+            dataDecrypted += (int) ((data[i] - key) % NUM_CHARS);
         }
         return dataDecrypted;
     }
